fix: resolve remaining compilation errors
- Add rusqlite::OptionalExtension import for .optional() method
- Fix seq type mismatch: cast i64 to u64 for wire protocol sequence
- Fix tokio read_exact returns Result<usize> not Result<()>
- Add Verifier trait import for ed25519_dalek::VerifyingKey::verify
- Fix not-Send errors: move .await calls before DB lock in store.rs
  (load_messages, get_message_count, mark_messages_read_for_peer,
  load_file_transfers, delete_all_data)

+++ src-tauri/src/messaging/mod.rs
@@ -165,7 +165,7 @@ pub async fn send_message(
         WireMessageType::Text,
         payload_bytes,
         msg_id.clone(),
-        seq,
+        seq as u64,
     )?;
     let wire_bytes = serialize_wire_message(&wire_msg)?;
